requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
billiard==3.6.0.0
celery==4.3.0
cffi==1.12.3
cryptography==2.6.1
Django==2.2.1
django-rest-knox==4.0.1
djangorestframework==3.9.3
kombu==4.5.0
pycparser==2.19
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
